small fix suggested by Hubert
On 14.09.2013, at 12:52, Hubert Wagner <hub.wag@gmail.com> wrote:

Hi Uli,

I am using PHAT for some experiments and it didn't compile under VC++ 2010. The problem was that std::back_inserter was not found in a number of cases. The solution is to  #include <iterator> in helpers/misc.h.

Best,
Hubert

git-svn-id: https://phat.googlecode.com/svn/trunk@143 8e3bb3c2-eed4-f18f-5264-0b6c94e6926d